+++ uikit/src/assets/styles/index.ts
@@ -1,6 +1,4 @@
-// @@TODO Build .css is 6MB now
-// When you remove that. Make sure you don't break playwright or storybook
-export {} from '@mdi/font/css/materialdesignicons.css';
+export {} from '@fortawesome/fontawesome-free/css/all.min.css';
 
 export {} from './_vars.scss';
 export {} from './styles.scss';