+++ uikit/src/components/KIcon/types.ts
@@ -1,7 +1,7 @@
 import type { KIconNames } from './iconNames';
 
 export type KIconProps = {
-  /** {@link https://pictogrammers.com/library/mdi/ pictogrammers.com} */
+  /** {@link https://fontawesome.com/icons fontawesome.com} */
   name: KIconNames;
   size?: number;
 };