+++ uikit/src/components/KTitle/__stories__/KTitle.stories.tsx
@@ -15,19 +15,19 @@ export const Primary: Story = {
     children: 'Title',
     actions: [
       {
-        iconLeft: 'account-box',
+        iconLeft: 'user',
         onClick: () => fn,
       },
       {
-        iconLeft: 'account-box',
+        iconLeft: 'user',
         onClick: () => fn,
       },
       {
-        iconLeft: 'account-box',
+        iconLeft: 'user',
         onClick: () => fn,
       },
       {
-        iconLeft: 'account-box',
+        iconLeft: 'user',
         onClick: () => fn,
       },
     ],