+++ uikit/src/components/KTitle/types.ts
@@ -5,4 +5,5 @@ export type KTitleAction = Pick<KButtonProps, 'iconLeft' | 'onClick'>;
 export type KTitleProps = {
   children?: React.ReactNode;
   actions?: KTitleAction[];
+  style?: React.CSSProperties;
 };